+++ CHANGELOG.md
@@ -22,6 +22,7 @@ all of them.
  - Add extension `VK_KHR_external_semaphore_win32` (revision 1)
  - Add extension `VK_KHR_external_semaphore_fd` (revision 1)
  - Add extension `VK_KHR_16bit_storage` (revision 1)
+ - Add extension `VK_KHR_external_fence_capabilities` (revision 1)
 
 ### Breaking changes